feat: adding exception handling to Player service

+++ src/main/java/com/luizmedeirosn/futs3/services/PlayerService.java
@@ -1,8 +1,11 @@
 package com.luizmedeirosn.futs3.services;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Service;
 
 import com.luizmedeirosn.futs3.dto.input.post.PostPlayerDTO;
@@ -18,6 +21,8 @@ import com.luizmedeirosn.futs3.repositories.ParameterRepository;
 import com.luizmedeirosn.futs3.repositories.PlayerParameterRepository;
 import com.luizmedeirosn.futs3.repositories.PlayerRepository;
 import com.luizmedeirosn.futs3.repositories.PositionRepository;
+import com.luizmedeirosn.futs3.services.exceptions.DatabaseException;
+import com.luizmedeirosn.futs3.services.exceptions.EntityNotFoundException;
 
 @Service
 public class PlayerService {
@@ -39,12 +44,22 @@ public class PlayerService {
     }
 
     public PlayerMinDTO findById(Long id) {
-        return new PlayerMinDTO( playerRepository.findById(id).get());
+        try {
+            return new PlayerMinDTO( playerRepository.findById(id).get());
+        
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("Player ID not found");
+        }
     }
 
 
     public PlayerDTO findByIdWithParameters(Long id) {
-        return new PlayerDTO( playerRepository.findById(id).get(), parameterRepository.findByPlayerId(id) );
+        try {
+            return new PlayerDTO( playerRepository.findById(id).get(), parameterRepository.findByPlayerId(id) );
+        
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("Player ID not found");
+        }
     }
 
     public List<AllPlayersParametersProjection> findAllWithParameters() {
@@ -52,30 +67,52 @@ public class PlayerService {
     }
 
     public PlayerDTO save(PostPlayerDTO playerInputDTO) {
-        Player newPlayer = new Player();
-        newPlayer.setName(playerInputDTO.getName());
-        newPlayer.setPosition( positionRepository.findById(playerInputDTO.getPositionId()).get() );
-        playerRepository.save(newPlayer);
+        try {
+            Player newPlayer = new Player();
+            newPlayer.setName(playerInputDTO.getName());
+            newPlayer.setPosition( positionRepository.findById(playerInputDTO.getPositionId()).get() );
+            playerRepository.save(newPlayer);
 
-        playerInputDTO.getParameters().forEach (
-            parameterScore -> {
-                Parameter parameter = parameterRepository.findById(parameterScore.getId()).get();
-                PlayerParameter playerParameter = 
-                    new PlayerParameter( newPlayer, parameter, parameterScore.getPlayerScore() );
-                playerParameterRepository.save(playerParameter);
-            }
-        );
-        return new PlayerDTO( newPlayer, parameterRepository.findByPlayerId(newPlayer.getId()) );
+            playerInputDTO.getParameters().forEach (
+                parameterScore -> {
+                    Parameter parameter = parameterRepository.findById(parameterScore.getId()).get();
+                    PlayerParameter playerParameter = 
+                        new PlayerParameter( newPlayer, parameter, parameterScore.getPlayerScore() );
+                    playerParameterRepository.save(playerParameter);
+                }
+            );
+            return new PlayerDTO( newPlayer, parameterRepository.findByPlayerId(newPlayer.getId()) );
+
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("Player request. IDs not found");
+
+        } catch (InvalidDataAccessApiUsageException e) {
+            throw new EntityNotFoundException("Player request. The given IDs must not be null");
+        
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Player request. Unique index, check index or primary key violation");
+        }
     }
 
     public PlayerMinDTO update(Long id, UpdatePlayerDTO updatePlayerDTO) {
-        Player player = playerRepository.getReferenceById(id);
-        player.updateData( updatePlayerDTO.getName(), positionRepository.findById(updatePlayerDTO.getPositionId()).get() );
-        player = playerRepository.save(player);
-        return new PlayerMinDTO(player);
+        try {
+            Player player = playerRepository.getReferenceById(id);
+            player.updateData( updatePlayerDTO.getName(), positionRepository.findById(updatePlayerDTO.getPositionId()).get() );
+            player = playerRepository.save(player);
+            return new PlayerMinDTO(player);
+
+        } catch (jakarta.persistence.EntityNotFoundException e) {
+            throw new EntityNotFoundException("Player request. IDs not found");
+        
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Player request. Unique index, check index or primary key violation");
+        }
     }
 
     public void deleteById(Long id) {
+        if (!playerRepository.existsById(id)) {
+            throw new EntityNotFoundException("Player request. ID not found");
+        }
         playerRepository.deleteById(id);
     }
